+++ frontend/apis/address.ts
@@ -1,29 +1,22 @@
-import { fetcher } from "./fetcher";
+import { api } from "./base";
 
-import { buildUrl } from "@/lib/utils";
 import { PaginatedAddress, Address } from "@/schemas";
 import { revalidate } from "@/actions/revalidate";
-import { ApiResult, tryCatch } from "@/lib/try-catch";
+import { ApiResult } from "@/lib/try-catch";
 import { Message } from "@/schemas";
 
 // Address API methods
 export const addressApi = {
     async all(input?: { search?: string; page?: number; limit?: number }): ApiResult<PaginatedAddress> {
         const searchParams = { search: input?.search || "", page: input?.page || 1, limit: input?.limit || 20 };
-        const url = buildUrl(`${process.env.NEXT_PUBLIC_BACKEND_URL}/api/address/`, searchParams);
-        const response = await tryCatch<PaginatedAddress>(fetcher(url, { next: { tags: ["addresses"] } }));
 
-        return response;
+        return await api.get<PaginatedAddress>("/address/", { params: { ...searchParams }, cache: "default", next: { tags: ["addresses"] } });
     },
     async get(id: number): ApiResult<Address> {
-        const url = `${process.env.NEXT_PUBLIC_BACKEND_URL}/api/address/${id}`;
-        const response = await tryCatch<Address>(fetcher(url));
-
-        return response;
+        return await api.get<Address>(`/address/${id}`);
     },
     async create(input: any): ApiResult<Address> {
-        const url = `${process.env.NEXT_PUBLIC_BACKEND_URL}/api/address/`;
-        const response = await tryCatch<Address>(fetcher(url, { method: "POST", body: JSON.stringify(input) }));
+        const response = await api.post<Address>(`/address/`, input);
 
         if (!response.error) {
             revalidate("addresses");
@@ -32,30 +25,16 @@ export const addressApi = {
         return response;
     },
     async update(id: number, input: any): ApiResult<Address> {
-        const url = `${process.env.NEXT_PUBLIC_BACKEND_URL}/api/address/${id}`;
-        const response = await tryCatch<Address>(fetcher(url, { method: "PATCH", body: JSON.stringify(input) }));
+        const response = await api.patch<Address>(`/address/${id}`, input);
 
         if (!response.error) {
             revalidate("addresses");
-            revalidate("cart");
-            revalidate("orders");
         }
 
         return response;
     },
     async delete(id: number): ApiResult<Message> {
-        const url = `${process.env.NEXT_PUBLIC_BACKEND_URL}/api/address/${id}`;
-        const response = await tryCatch<Message>(fetcher(url, { method: "DELETE" }));
-
-        if (!response.error) {
-            revalidate("addresses");
-        }
-
-        return response;
-    },
-    async billing(input: any): ApiResult<Address> {
-        const url = `${process.env.NEXT_PUBLIC_BACKEND_URL}/api/address/billing`;
-        const response = await tryCatch<Address>(fetcher(url, { method: "POST", body: JSON.stringify(input) }));
+        const response = await api.delete<Message>(`/address/${id}`);
 
         if (!response.error) {
             revalidate("addresses");
