+++ frontend/app/(main)/(static)/terms/page.tsx
@@ -59,7 +59,7 @@ const Terms = async () => {
             </div>
 
             <div className="mt-8 text-center">
-                <BtnLink color="primary" href="/" size="md">
+                <BtnLink href="/" size="lg" variant="primary">
                     Back Home
                 </BtnLink>
             </div>
