+++ frontend/app/(main)/(static)/user-agreement/page.tsx
@@ -55,7 +55,7 @@ const UserAgreement = async () => {
             </div>
 
             <div className="mt-8 text-center">
-                <BtnLink color="primary" href="/" size="md">
+                <BtnLink href="/" size="lg" variant="primary">
                     Back Home
                 </BtnLink>
             </div>
