+++ frontend/schemas/index.ts
@@ -1,12 +1,13 @@
 export * from "./enums";
 export * from "./common";
-export * from "./user";
 export * from "./address";
+export * from "./user";
 export * from "./product";
 export * from "./cart";
 export * from "./order";
-export * from "./pagination";
 export * from "./review";
+export * from "./activity";
+export * from "./conversation";
 
 export interface DeliveryOption {
     id: number;
